requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.1.1
django-crispy-forms==2.3
djangorestframework==3.15.2
idna==3.10
pillow==10.4.0
<<<<<<< HEAD
psycopg2-binary==2.9.9
=======
>>>>>>> 1500c71627aee91dfad6bb07462a2105c050169d
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
